backend/pyproject.toml
[project]
name = "harvey-backend"
version = "0.1.0"
description = "FastAPI skeleton for BMW contract review assistant"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110.0",
  "httpx>=0.27.0",
  "pydantic-settings>=2.2.1",
  "uvicorn[standard]>=0.27.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
